git-svn-id: http://svn.encuesta.me/trunk/encuestame@315 39e2479b-dd58-4041-9f67-69c9b71f0610

+++ source/java/org/encuestame/core/persistence/dao/imp/ICatLocation.java
@@ -51,7 +51,7 @@ public interface ICatLocation extends IBaseDao {
      * @return aa
       * @throws HibernateException HibernateException
      */
-    public List<CatLocation> getLocationbyLevelId(final Long locateId) throws HibernateException;
+    public List<CatLocation> getLocationbyLevelId(final Integer locateId) throws HibernateException;
 
     /**
      * @param tidtype tidtype
